% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfox_to_move2.R
\name{mt_preprocess}
\alias{mt_preprocess}
\title{Preprocess move2 Data}
\usage{
mt_preprocess(m2)
}
\arguments{
\item{m2}{move2 object.}
}
\value{
Preprocessed move2 object.
}
\description{
Ensures move2 data is clean and ready for analysis.
}
