% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfox_download.R
\name{process_data}
\alias{process_data}
\title{Process downloaded data}
\usage{
process_data(df, capture_data)
}
\arguments{
\item{df}{Data frame of downloaded data.}

\item{capture_data}{Data frame of capture information.}
}
\value{
A data frame with combined and processed information.
}
\description{
Process downloaded data
}
