% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfox_download.R
\name{sigfox_download}
\alias{sigfox_download}
\title{Download Sigfox Tracking Data}
\usage{
sigfox_download(
  tag_ID = NA,
  tag_type = "tinyfox",
  ID = NA,
  ring = NA,
  attachment_type = NA,
  capture_weight = NA,
  capture_time = NA,
  FA_length = NA,
  tag_weight = NA,
  sex = NA,
  age = NA,
  repro_status = NA,
  species = NA,
  release_time = NA,
  capture_latitude = NA,
  capture_longitude = NA,
  deploy_on_latitude = NA,
  deploy_on_longitude = NA,
  roost = NA,
  firmware = NA,
  download_attempts = 5
)
}
\arguments{
\item{tag_ID}{Character; Biologger ID.}

\item{tag_type}{Character; Tag version, e.g. tinyfox, nanofox, torpor}

\item{ID}{Character; PIT-tag identifier.}

\item{ring}{Character; Ring identifier.}

\item{attachment_type}{Character; Type of attachment, e.g., 'glue' or 'collar'.}

\item{capture_weight}{Numeric; Weight of the bat at capture in grams.}

\item{capture_time}{POSIXct; DateTime of capture.}

\item{FA_length}{Numeric; Forearm length in mm.}

\item{tag_weight}{Numeric; Weight of the tag in grams.}

\item{sex}{Character; Sex of the bat.}

\item{age}{Character; Age of the bat.}

\item{repro_status}{Character; Reproductive status of the bat.}

\item{species}{Character; Species of the bat.}

\item{release_time}{POSIXct; DateTime of release.}

\item{capture_latitude}{Numeric; Latitude of capture location.}

\item{capture_longitude}{Numeric; Longitude of capture location.}

\item{roost}{Character; Roost identifier.}

\item{firmware}{Character; Firmware version of the tag}

\item{download_attempts}{Numeric; Number of times to retry downloads of the sigfox data}
}
\value{
A data frame of the deployment data including tracking and bat information.
}
\description{
This function downloads tracking data for individual tags from a specified source.
It retries the download multiple times in case of connection failure.
}
\examples{
\dontrun{
  downloaded_data <- sigfox_download(ID = "12345", tag_ID = "ABC123", ...)
}
}
